apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transactional-control-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "python-jose[cryptography]>=3.3",
  "bcrypt>=4.1",
  "celery>=5.3",
  "redis>=5.0",
  "httpx>=0.27",
  "pytest>=8.0",
  "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tc"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["src/tc/db/migrations/versions", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["src/tc/tests"]
